Many tests fixes

+++ src/main/java/it/attsd/deepsky/model/ConstellationRepository.java
@@ -3,37 +3,41 @@ package it.attsd.deepsky.model;
 import java.util.List;
 
 import javax.persistence.NoResultException;
+import javax.persistence.PersistenceException;
 import javax.persistence.Query;
 import javax.transaction.Transactional;
 
+import org.apache.commons.lang3.exception.ExceptionUtils;
+import org.hibernate.exception.ConstraintViolationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
 import it.attsd.deepsky.entity.Constellation;
-import it.attsd.deepsky.exception.RepositoryException;
+import it.attsd.deepsky.exception.ConstellationAlreadyExistsException;
+import it.attsd.deepsky.exception.GenericRepositoryException;
 
 @Repository
 public class ConstellationRepository extends BaseRepository {
 	private Logger logger = LoggerFactory.getLogger(ConstellationRepository.class);
-	
+
 	@Transactional
 	public void emptyTable() {
 		entityManager.createQuery(String.format("DELETE FROM %s", Constellation.class.getName())).executeUpdate();
 	}
-	
+
 	@SuppressWarnings("unchecked")
 	public List<Constellation> findAll() {
 		Query query = entityManager.createQuery(String.format("SELECT t FROM %s t", Constellation.class.getName()));
 
-		return (List<Constellation>) query.getResultList();
+		return query.getResultList();
 	}
 
 	public Constellation findById(long id) {
 		return entityManager.find(Constellation.class, id);
 	}
 
-	public Constellation findByName(String name) throws RepositoryException {
+	public Constellation findByName(String name) {
 		Constellation result = null;
 		try {
 			Query query = entityManager
@@ -42,47 +46,43 @@ public class ConstellationRepository extends BaseRepository {
 			result = (Constellation) query.getSingleResult();
 		} catch (NoResultException e) {
 			logger.info(String.format("No Constellation found with name %s", name));
-		} catch (Exception e) {
-			logger.error(e.getMessage());
-			throw new RepositoryException(e);
 		}
 		return result;
 	}
 
 	@Transactional
-	public Constellation save(Constellation constellation) throws RepositoryException {
+	public Constellation save(Constellation constellation)
+			throws GenericRepositoryException, ConstellationAlreadyExistsException {
 		try {
 			entityManager.persist(constellation);
 			entityManager.flush();
+		} catch (PersistenceException e) {
+			if (ExceptionUtils.indexOfType(e, ConstraintViolationException.class) != -1) {
+				throw new ConstellationAlreadyExistsException();
+			} else {
+				throw e;
+			}
 		} catch (Exception e) {
-			logger.error(e.getMessage());
-			throw new RepositoryException(e);
+			throw new GenericRepositoryException(e);
 		}
 
 		return constellation;
 	}
-	
+
 	@Transactional
-	public void update(Constellation constellation) throws RepositoryException {
-		try {
-			entityManager.merge(constellation);
-			entityManager.flush();
-		} catch (Exception e) {
-			logger.error(e.getMessage());
-			throw new RepositoryException(e);
-		}
+	public Constellation update(Constellation constellation) {
+		Constellation constellationChanged = entityManager.merge(constellation);
+		entityManager.flush();
+		
+		return constellationChanged;
 	}
 
 	@Transactional
-	public void delete(long id) throws RepositoryException {
-		try {
-			Constellation constellation = findById(id);
-
+	public void delete(long id) {
+		Constellation constellation = findById(id);
+		if(constellation != null) {
 			entityManager.remove(constellation);
 			entityManager.flush();
-		} catch (Exception e) {
-			logger.error(e.getMessage());
-			throw new RepositoryException(e);
 		}
 	}
