+++ src/main/java/it/attsd/deepsky/model/DeepSkyObjectRepository.java
@@ -3,38 +3,41 @@ package it.attsd.deepsky.model;
 import java.util.List;
 
 import javax.persistence.NoResultException;
+import javax.persistence.PersistenceException;
 import javax.persistence.Query;
 import javax.transaction.Transactional;
 
+import org.apache.commons.lang3.exception.ExceptionUtils;
+import org.hibernate.exception.ConstraintViolationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
 import it.attsd.deepsky.entity.DeepSkyObject;
-import it.attsd.deepsky.entity.DeepSkyObjectType;
-import it.attsd.deepsky.exception.RepositoryException;
+import it.attsd.deepsky.exception.DeepSkyObjectAlreadyExistsException;
+import it.attsd.deepsky.exception.GenericRepositoryException;
 
 @Repository
 public class DeepSkyObjectRepository extends BaseRepository {
 	private Logger logger = LoggerFactory.getLogger(DeepSkyObjectRepository.class);
-	
+
 	@Transactional
 	public void emptyTable() {
 		entityManager.createQuery(String.format("DELETE FROM %s", DeepSkyObject.class.getName())).executeUpdate();
 	}
-	
+
 	@SuppressWarnings("unchecked")
 	public List<DeepSkyObject> findAll() {
 		Query query = entityManager.createQuery(String.format("SELECT t FROM %s t", DeepSkyObject.class.getName()));
 
-		return (List<DeepSkyObject>) query.getResultList();
+		return query.getResultList();
 	}
 
 	public DeepSkyObject findById(long id) {
 		return entityManager.find(DeepSkyObject.class, id);
 	}
 
-	public DeepSkyObject findByName(String name) throws RepositoryException {
+	public DeepSkyObject findByName(String name) {
 		DeepSkyObject result = null;
 		try {
 			Query query = entityManager
@@ -43,47 +46,42 @@ public class DeepSkyObjectRepository extends BaseRepository {
 			result = (DeepSkyObject) query.getSingleResult();
 		} catch (NoResultException e) {
 			logger.info(String.format("No DeepSkyObject found with name %s", name));
-		} catch (Exception e) {
-			logger.error(e.getMessage());
-			throw new RepositoryException(e);
 		}
 		return result;
 	}
 
 	@Transactional
-	public DeepSkyObject save(DeepSkyObject deepSkyObject) throws RepositoryException {
+	public DeepSkyObject save(DeepSkyObject deepSkyObject) throws GenericRepositoryException, DeepSkyObjectAlreadyExistsException {
 		try {
 			entityManager.persist(deepSkyObject);
 			entityManager.flush();
+		} catch (PersistenceException e) {
+			if (ExceptionUtils.indexOfType(e, ConstraintViolationException.class) != -1) {
+				throw new DeepSkyObjectAlreadyExistsException();
+			} else {
+				throw e;
+			}
 		} catch (Exception e) {
-			logger.error(e.getMessage());
-			throw new RepositoryException(e);
+			throw new GenericRepositoryException(e);
 		}
 
 		return deepSkyObject;
 	}
-	
+
 	@Transactional
-	public void update(DeepSkyObject deepSkyObject) throws RepositoryException {
-		try {
-			entityManager.merge(deepSkyObject);
-			entityManager.flush();
-		} catch (Exception e) {
-//			logger.error(e.getMessage());
-			throw new RepositoryException(e);
-		}
+	public DeepSkyObject update(DeepSkyObject deepSkyObject) {
+		DeepSkyObject deepSkyObjectChanged = entityManager.merge(deepSkyObject);
+		entityManager.flush();
+
+		return deepSkyObjectChanged;
 	}
 
 	@Transactional
-	public void delete(long id) throws RepositoryException {
-		try {
-			DeepSkyObject deepSkyObject = findById(id);
-
+	public void delete(long id) {
+		DeepSkyObject deepSkyObject = findById(id);
+		if (deepSkyObject != null) {
 			entityManager.remove(deepSkyObject);
 			entityManager.flush();
-		} catch (Exception e) {
-			logger.error(e.getMessage());
-			throw new RepositoryException(e);
 		}
 	}
