+++ src/main/java/it/attsd/deepsky/model/DeepSkyObjectTypeRepository.java
@@ -3,16 +3,23 @@ package it.attsd.deepsky.model;
 import java.util.List;
 
 import javax.persistence.NoResultException;
+import javax.persistence.PersistenceException;
 import javax.persistence.Query;
 import javax.transaction.Transactional;
 
+import org.apache.commons.lang3.exception.ExceptionUtils;
+import org.hibernate.exception.ConstraintViolationException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
 import it.attsd.deepsky.entity.DeepSkyObjectType;
-import it.attsd.deepsky.exception.RepositoryException;
+import it.attsd.deepsky.exception.DeepSkyObjectTypeAlreadyExistsException;
+import it.attsd.deepsky.exception.GenericRepositoryException;
 
 @Repository
 public class DeepSkyObjectTypeRepository extends BaseRepository {
+	private Logger logger = LoggerFactory.getLogger(DeepSkyObjectTypeRepository.class);
 
 	@Transactional
 	public void emptyTable() {
@@ -23,14 +30,14 @@ public class DeepSkyObjectTypeRepository extends BaseRepository {
 	public List<DeepSkyObjectType> findAll() {
 		Query query = entityManager.createQuery(String.format("SELECT t FROM %s t", DeepSkyObjectType.class.getName()));
 
-		return (List<DeepSkyObjectType>) query.getResultList();
+		return query.getResultList();
 	}
 
 	public DeepSkyObjectType findById(long id) {
 		return entityManager.find(DeepSkyObjectType.class, id);
 	}
 
-	public DeepSkyObjectType findByType(String type) throws RepositoryException {
+	public DeepSkyObjectType findByType(String type) {
 		DeepSkyObjectType result = null;
 		try {
 			Query query = entityManager.createQuery(
@@ -38,45 +45,43 @@ public class DeepSkyObjectTypeRepository extends BaseRepository {
 			query.setParameter("type", type.toLowerCase());
 			result = (DeepSkyObjectType) query.getSingleResult();
 		} catch (NoResultException e) {
-
-		} catch (Exception e) {
-			throw new RepositoryException(e);
+			logger.info(String.format("No DeepSkyObjectType found with type %s", type));
 		}
 		return result;
 	}
 
 	@Transactional
-	public DeepSkyObjectType save(DeepSkyObjectType deepSkyObjectType) throws RepositoryException {
+	public DeepSkyObjectType save(DeepSkyObjectType deepSkyObjectType) throws GenericRepositoryException, DeepSkyObjectTypeAlreadyExistsException {
 		try {
 			entityManager.persist(deepSkyObjectType);
 			entityManager.flush();
+		} catch (PersistenceException e) {
+			if (ExceptionUtils.indexOfType(e, ConstraintViolationException.class) != -1) {
+				throw new DeepSkyObjectTypeAlreadyExistsException();
+			} else {
+				throw e;
+			}
 		} catch (Exception e) {
-			throw new RepositoryException(e);
+			throw new GenericRepositoryException(e);
 		}
 
 		return deepSkyObjectType;
 	}
-	
+
 	@Transactional
-	public void update(DeepSkyObjectType deepSkyObjectType) throws RepositoryException {
-		try {
-			entityManager.merge(deepSkyObjectType);
-			entityManager.flush();
-		} catch (Exception e) {
-//			logger.error(e.getMessage());
-			throw new RepositoryException(e);
-		}
+	public DeepSkyObjectType update(DeepSkyObjectType deepSkyObjectType) {
+		DeepSkyObjectType deepSkyObjectTypeChanged = entityManager.merge(deepSkyObjectType);
+		entityManager.flush();
+
+		return deepSkyObjectTypeChanged;
 	}
 
 	@Transactional
-	public void delete(long id) throws RepositoryException {
-		try {
-			DeepSkyObjectType deepSkyObjectType = findById(id);
-
+	public void delete(long id) {
+		DeepSkyObjectType deepSkyObjectType = findById(id);
+		if (deepSkyObjectType != null) {
 			entityManager.remove(deepSkyObjectType);
 			entityManager.flush();
-		} catch (Exception e) {
-			throw new RepositoryException(e);
 		}
 	}
